+++ include/fluxgl/ecs/components/audio_listener.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace fluxgl {
+    struct AudioListener {
+        bool active = true;
+    };
+}